+++ src/main/java/be/jessevdk/CryptoSimulator/models/dto/PortfolioDTO.java
@@ -2,16 +2,17 @@ package be.jessevdk.CryptoSimulator.models.dto;
 
 import be.jessevdk.CryptoSimulator.models.domain.Coin;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class PortfolioDTO {
     private List<CoinDTO> coins;
-    private double valueUsd; //todo Floating-Point Arithmetic, double should not be used for financial precise calculations
+    private BigDecimal valueUsd; //todo Floating-Point Arithmetic, double should not be used for financial precise calculations
 
     public PortfolioDTO() {
     }
 
-    public PortfolioDTO(List<CoinDTO> coins, double valueUsd) {
+    public PortfolioDTO(List<CoinDTO> coins, BigDecimal valueUsd) {
         this.coins = coins;
         this.valueUsd = valueUsd;
     }
@@ -24,11 +25,11 @@ public class PortfolioDTO {
         this.coins = coins;
     }
 
-    public double getValueUsd() {
+    public BigDecimal getValueUsd() {
         return valueUsd;
     }
 
-    public void setValueUsd(double valueUsd) {
+    public void setValueUsd(BigDecimal valueUsd) {
         this.valueUsd = valueUsd;
     }
 }
